shop2/requirements.txt
flask
sqlalchemy
waitress
psycopg
azure-identity
azure-keyvault-secrets
